+++ data-raw/data-internal.R
@@ -1,0 +1,6 @@
+rm(list = ls())
+
+load("data-raw/map.rda")
+
+use_data(map, pkg = as.package("."), internal = TRUE,
+         overwrite = TRUE, compress = "xz")