+++ data-raw/data-raw.R
@@ -1,10 +1,11 @@
-#source("data-raw/fraser-raw.R")
-#source("data-raw/map-raw.R")
-
 library(wqbc) # rebuild package
+source("data-raw/dummy.R")
 source("data-raw/codes.R")
 source("data-raw/limits.R")
 library(wqbc) # rebuild package
+# source("data-raw/fraser-raw.R")
 source("data-raw/fraser.R")
+# source("data-raw/map-raw.R")
 source("data-raw/map.R")
 source("data-raw/ccme.R")
+source("data-raw/data-internal.R")