+++ demo/fraser.R
@@ -4,20 +4,27 @@ library(ggplot2)
 library(sp)
 library(rgdal)
 
-data(fraser)
+options(wqbc.messages = TRUE)
 
-fraser$Variable <- substitute_variables(fraser$Variable)
-fraser$Units <- substitute_units(fraser$Units)
-fraser <- filter(fraser, Variable %in% get_variables() & Units %in% get_units())
+data(fraser)
+print(summary(fraser))
+
 fraser$SiteID <-  factor(sub("BC08", "", as.character(fraser$SiteID)))
+fraser$Year <- year(fraser$Date)
 
 plot_map(fraser, fill = "SiteID")
 
-fraser_limits <- calc_limits(fraser, by = c("SiteID", "Lat", "Long"))
+fraser <- calc_wqis(fraser, by = c("SiteID", "Year", "Lat", "Long"))
 
-fraser_limits$Year <- year(fraser_limits$Date)
-fraser_wqis <- calc_wqis(fraser_limits, by = c("SiteID", "Year", "Lat", "Long"))
+plot_wqis(fraser, x = "Year") + facet_wrap(~SiteID)
+plot_map_wqis(filter(fraser, Year %in% c(2000,2010)), keep = "Year") +
+  facet_wrap(~Year, ncol = 1)
 
-plot_wqis(fraser_wqis, x = "Year") + facet_wrap(~SiteID)
-
-plot_map_wqis(filter(fraser_wqis, Year %in% c(1990,2010)), keep = "Year") + facet_wrap(~Year, ncol = 1)
+# Short-term WQIs by year
+data(fraser)
+fraser$Year <- year(fraser$Date)
+fraser <- standardize_wqdata(fraser, strict = FALSE)
+fraser <- clean_wqdata(fraser, by = "Year", max_cv = Inf)
+fraser <- calc_limits(fraser, by = "Year", term = "short")
+fraser <- calc_wqis(fraser, by = "Year")
+plot_wqis(fraser, x = "Year")