+++ NEWS.md
@@ -1,6 +1,10 @@
-# wqbc 0.0.0.9001
+# wqbc 0.0.0.9005
 
-- No changes to date.
+- Released new version number as wqdata now dependent
+
+# wqbc 0.0.0.9002
+
+- Finalized limits.csv and codes.csv files
 
 # wqbc 0.0.0.9000